Extract QESSettings into its own class
In the process, add a seekbar, a play/pause button, and a toggle
(Spacebar) on whether the mouse-look script is active.  This means that
when you want to interact with the UI, you can press spacebar to toggle
whether moving the mouse looks around, you can manipulate the UI, and
you can press Spacebar again to go back to using the mouse look.